+++ packages/integration-tests/src/test-project-setup/data_storage_auth_with_triggers.ts
@@ -33,6 +33,10 @@ import {
 import { e2eToolingClientConfig } from '../e2e_tooling_client_config.js';
 import isMatch from 'lodash.ismatch';
 import { TextWriter, ZipReader } from '@zip.js/zip.js';
+import {
+  AdminCreateUserCommand,
+  CognitoIdentityProviderClient,
+} from '@aws-sdk/client-cognito-identity-provider';
 
 /**
  * Creates test projects with data, storage, and auth categories.
@@ -68,7 +72,10 @@ export class DataStorageAuthWithTriggerTestProjectCreator
     private readonly cloudTrailClient: CloudTrailClient = new CloudTrailClient(
       e2eToolingClientConfig
     ),
-    private readonly resourceFinder: DeployedResourcesFinder = new DeployedResourcesFinder()
+    private readonly resourceFinder: DeployedResourcesFinder = new DeployedResourcesFinder(),
+    private readonly cognitoClient: CognitoIdentityProviderClient = new CognitoIdentityProviderClient(
+      e2eToolingClientConfig
+    )
   ) {}
 
   createProject = async (e2eProjectDir: string): Promise<TestProjectBase> => {
@@ -87,7 +94,8 @@ export class DataStorageAuthWithTriggerTestProjectCreator
       this.iamClient,
       this.sqsClient,
       this.cloudTrailClient,
-      this.resourceFinder
+      this.resourceFinder,
+      this.cognitoClient
     );
     await fs.cp(
       project.sourceProjectAmplifyDirURL,
@@ -154,7 +162,8 @@ class DataStorageAuthWithTriggerTestProject extends TestProjectBase {
     private readonly iamClient: IAMClient,
     private readonly sqsClient: SQSClient,
     private readonly cloudTrailClient: CloudTrailClient,
-    private readonly resourceFinder: DeployedResourcesFinder
+    private readonly resourceFinder: DeployedResourcesFinder,
+    private readonly cognitoClient: CognitoIdentityProviderClient
   ) {
     super(
       name,
@@ -262,12 +271,19 @@ class DataStorageAuthWithTriggerTestProject extends TestProjectBase {
       'AWS::Lambda::Function',
       (name) => name.includes('funcNoMinify')
     );
+    const funcCustomEmailSender =
+      await this.resourceFinder.findByBackendIdentifier(
+        backendId,
+        'AWS::Lambda::Function',
+        (name) => name.includes('funcCustomEmailSender')
+      );
 
     assert.equal(defaultNodeLambda.length, 1);
     assert.equal(node16Lambda.length, 1);
     assert.equal(funcWithSsm.length, 1);
     assert.equal(funcWithAwsSdk.length, 1);
     assert.equal(funcWithSchedule.length, 1);
+    assert.equal(funcCustomEmailSender.length, 1);
 
     const expectedResponse = {
       s3TestContent: 'this is some test content',
@@ -279,7 +295,9 @@ class DataStorageAuthWithTriggerTestProject extends TestProjectBase {
     await this.checkLambdaResponse(defaultNodeLambda[0], expectedResponse);
     await this.checkLambdaResponse(node16Lambda[0], expectedResponse);
     await this.checkLambdaResponse(funcWithSsm[0], 'It is working');
-    await this.checkLambdaResponse(funcWithAwsSdk[0], 'It is working');
+
+    // Custom email sender assertion
+    await this.assertCustomEmailSenderWorks(backendId);
 
     await this.assertScheduleInvokesFunction(backendId);
 
@@ -641,5 +659,84 @@ class DataStorageAuthWithTriggerTestProject extends TestProjectBase {
         `The scheduled function failed to invoke and send a message to the queue.`
       );
     }
+  };
+
+  private assertCustomEmailSenderWorks = async (
+    backendId: BackendIdentifier
+  ) => {
+    const TIMEOUT_MS = 1000 * 60 * 2; // 2 minutes
+    const startTime = Date.now();
+    const queue = await this.resourceFinder.findByBackendIdentifier(
+      backendId,
+      'AWS::SQS::Queue',
+      (name) => name.includes('customEmailSenderQueue')
+    );
+
+    assert.strictEqual(queue.length, 1, 'Custom email sender queue not found');
+
+    // Trigger an email sending operation
+    await this.triggerEmailSending(backendId);
+
+    // Wait for the SQS message
+    let messageReceived = false;
+    while (Date.now() - startTime < TIMEOUT_MS && !messageReceived) {
+      const response = await this.sqsClient.send(
+        new ReceiveMessageCommand({
+          QueueUrl: queue[0],
+          WaitTimeSeconds: 20,
+        })
+      );
+
+      if (response.Messages && response.Messages.length > 0) {
+        messageReceived = true;
+        // Verify the message content
+        const messageBody = JSON.parse(response.Messages[0].Body || '{}');
+        assert.strictEqual(
+          messageBody.message,
+          'Custom Email Sender is working',
+          'Unexpected message content'
+        );
+
+        // Delete the message
+        await this.sqsClient.send(
+          new DeleteMessageCommand({
+            QueueUrl: queue[0],
+            ReceiptHandle: response.Messages[0].ReceiptHandle!,
+          })
+        );
+      }
+    }
+
+    assert.strictEqual(
+      messageReceived,
+      true,
+      'Custom email sender was not triggered within the timeout period'
+    );
+  };
+
+  private triggerEmailSending = async (backendId: BackendIdentifier) => {
+    const userPoolId = await this.resourceFinder.findByBackendIdentifier(
+      backendId,
+      'AWS::Cognito::UserPool',
+      () => true
+    );
+
+    assert.strictEqual(userPoolId.length, 1, 'User pool not found');
+
+    const username = `testuser_${Date.now()}@example.com`;
+    const password = 'TestPassword123!';
+
+    await this.cognitoClient.send(
+      new AdminCreateUserCommand({
+        UserPoolId: userPoolId[0],
+        Username: username,
+        TemporaryPassword: password,
+        UserAttributes: [
+          { Name: 'email', Value: username },
+          { Name: 'email_verified', Value: 'true' },
+        ],
+      })
+    );
+    // The creation of a new user should trigger the custom email sender
   };
 }