+++ packages/backend-auth/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @aws-amplify/backend-auth
 
+## 1.3.0
+
+### Minor Changes
+
+- 11d62fe: Add support for custom Lambda function email senders in Auth construct
+
+### Patch Changes
+
+- b56d344: update aws-cdk lib to ^2.158.0
+- Updated dependencies [11d62fe]
+- Updated dependencies [b56d344]
+  - @aws-amplify/auth-construct@1.4.0
+  - @aws-amplify/backend-output-storage@1.1.3
+  - @aws-amplify/plugin-types@1.3.1
+
 ## 1.2.0
 
 ### Minor Changes