+++ packages/backend-platform-test-stubs/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @aws-amplify/backend-platform-test-stubs
 
+## 0.3.6
+
+### Patch Changes
+
+- b56d344: update aws-cdk lib to ^2.158.0
+- Updated dependencies [b56d344]
+  - @aws-amplify/plugin-types@1.3.1
+
 ## 0.3.5
 
 ### Patch Changes